+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Thu, 23 Jun 2022 16:08:07 GMT and should not be manually modified.
+This log was last generated on Fri, 24 Jun 2022 22:38:56 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.68.3
+
+Fri, 24 Jun 2022 22:38:56 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu to v0.15.6
+
+## 0.68.2
+
+Fri, 24 Jun 2022 18:52:57 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/win32-theme to v0.19.1
+
+## 0.68.1
+
+Thu, 23 Jun 2022 22:10:22 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu to v0.15.5
 
 ## 0.68.0